+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Once we have a solid base, we will bump to v1 without any breaking changes from 
 * Ability to use custom names for your export functions
 * NXPorts will automatically remove the Prefer32Bit flag and change it to require 32bit
 * NXPorts can be added via NuGet package and the export reweave process happens on build
+* Attributes used to mark the desired exports will now be removed, thus removing NXPorts.Attributes.dll from the assembly dependencies.